sandbox/requirements.txt
apache-beam[gcp]<=2.67.0
google-cloud==3.4.0
paho-mqtt==2.1.0
locust==2.39.0